+++ src/app/data/poe-prices/index.ts
@@ -1,0 +1,1 @@
+export * from './service/item-price-prediction-http.service';