+++ src/app/shared/module/poe/service/index.ts
@@ -2,6 +2,7 @@ export * from './context.service';
 export * from './item/item-clipboard.service';
 export * from './item/item-exchange-rate.service';
 export * from './item/item-external.service';
+export * from './item/item-price-prediction.service';
 export * from './item/item-search-analyze.service';
 export * from './item/item-search.service';
 export * from './item/item.service';